[FIO toup] zynqmp: set modeboot unconditionally
If env is reset with env default -a, modeboot also becomes empty and
it is never set after as we hit this condition after each boot:

if (!(gd->flags & GD_FLG_ENV_DEFAULT)) {
    debug("Saved variables - Skipping\n");
    return 0;
}

This causes all logic, which depends on modeboot value, work
incorrectly. Set modeboot on every boot unconditionally.

Signed-off-by: Igor Opaniuk <igor.opaniuk@foundries.io>
Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>